Extract reducer and internal hook
